change duration to double
git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/InductiveMiner/Trunk@43791 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/processmining/plugins/inductiveminer2/attributes/Attribute.java
@@ -53,17 +53,17 @@ public interface Attribute extends Comparable<Attribute> {
 	 */
 	public long getTime(XAttributable x);
 
-	public long getDurationMin();
+	public double getDurationMin();
 
-	public long getDurationMax();
+	public double getDurationMax();
 
 	/**
-	 * The time value, or Long.MIN_VALUE if it does not exist.
+	 * The duration value, or -Double.MAX_VALUE if it does not exist.
 	 * 
 	 * @param x
 	 * @return
 	 */
-	public long getDuration(XAttributable x);
+	public double getDuration(XAttributable x);
 
 	public int compareTo(Attribute arg0);
